Update master to version 8.1 (#68)
* Revert "Temporarily disable  -warnings-as-errors, investigate whole-module build"

This reverts commit 80380d26aa351bd5ad518750d56463a9b18c4aeb.

* Fix build

* Fix build

* Bump version

* Fix more warnings

* Fix more warnings

* Fix some errors

* Fix build

* Fix build

* Fix build

* Fix more warnings

* Fix more warnings

* Refactoring

* Fix negative ids

* Add try-catch

* Combo update

* Various Improvements

* Fix videoplayer UI in landscape

* Fix trending packs shimmer animation

* Fix trending packs animation playback

* Fix static stickerpack thumbnail

* Various Fixes

* Animation updates

* Fix more warnings

* Add sticker packs reordering

* Change media saving option strings

* Fix audio playback rate

* Refactoring

* Various Improvements

* Various Improvements

* Support video streaming API (PoC)

* Fix decode warning

* Change quality

* Show trending icon in packs scroller when trending strip is dismissed in the grid

* Various theme color picking fixes

* Improved streaming

* Various theme editing imrpovements

* Add "Choosing sticker" input activity

* Various Fixes

* Various theme editing fixes

* Various theme editing fixes

* Theme decoding fixes

* Various theme editing fixes

* syntax fixes

* Fix theme wallpaper coding

* Theme editing fixes

* Update API and tgcalls

* Update tgcalls

* Update tgcalls

* Implemented chat-specific themes

* Temporary fi
x

* Chat-specific theme improvements

* Chat-specific theme improvements

* Initial ad messages implementation

* Fix ad buttons

* Update tgcalls

* Update API

* Various Improvements

* Fix pattern wallpaper rendering

* Fix cdn file random access

* Fix accessory panel icons

* Send choosing sticker activity

* Fix bubbles background occasionally not rendering first time

* Fix choosing sticker activity layout in chat list

* Various Fixes

* Update tgcalls

* Fix text

* Update tgcalls

* Bump version

* Fix game bot popup

* Update github actions

* Fix build

* Fix build

* Temporarily disable -Werror in LegacyComponents

* Channel UI fixes

* Various Fixes

* Fix input media panel icons

* Various Fixes

* Various Fixes

* Various Fixes

* UI updates

* Update tgcalls

* Various Fixes

* Various Fixes

* Fix crash

* Fix media selection

* Revert tgcalls

* Update tgcalls

* Update tgcalls

* Update tgcalls

* Various Fixes

* Various Fixes

* Various Fixes

* Various Fixes

* Various Fixes

* Various Fixes

* Separate channel voice chat localization

* Channel navigation improvements

* Revert localization [skip ci]

* Fix 32-bit crashes [skip ci]

* Various Fixes

* Update localization

* Various Fixes

* Various Fixes

* Update localization

* Various Fixes

* Revert voice chat recording UI

* Various Fixes

* Various Fixes

* Update localization

* Fix localization

* Move overscroll view to the parent view

* Fix choosing sticker activity for groups

* Fix trending packs dismissal

* Fix localization

* Fix build

* Various Fixes

* Various Fixes

* Fix Send Message option text in forwarding

* Fix trending packs dismissal

* Optimize pattern wallpapers rendering

* Chat-specific themes improvements

* Update tgcalls

* Update dark chat theme

* Various Improvements

* Support view list

* Chat-specific theme improvements

* Add outgoing accent color customization

* Test additional emoji animations

* Various theme editor fixes

* Update read list UI

* Various theme editor fixes

* Various theme fixes

* Theme editor fixes

* Various Improvements

* Theme editor improvements

* Various theme improvements

* UI fixes

* Update localization

* Various Improvements

* Fix backdrop animation

* Fix send button

* Fix instant video background not updating in theme change

* Fix dark appearance flag

* Fix read state sync

* Bump version

* Update animations

* Update animations

* Fix theme preview crossfade

* Update context menu animation

* Various chat theme fixes

* Update localization

* Fix cache

* Fix typo

* Add animated emojis in chat theme selection screen

* Fix chat theme preview

* Various Fixes

* Chat theme fixes

* Context UI animations

* Update Context UI

* Emoji animations improvements

* Various Fixes

* Fix context UI

* Various chat theme fixes

* Periodically update avialable chat themes list

* Fix string fallback

* UI fixes

* Adjust overscroll dismiss

* Various fixes

* Update localization

* Various Fixes

* Fix urlencode

* Fix context UI positioning

* Fix synthetic user id hashes

* Update API

* Update emoji animation time offset rounding

* Fix emoji interaction json parsing

* Animated emoji animations fixes

* Fix color

* Fix url parsing

* Various Improvements

* Play haptic feedback when receiving emoji interaction

* Return voice chat recording title setup

* Update emoji animations haptic

* Fix chat theme tooltip deactivation

* Bump version

* Various Fixes

* Set choosing sticker activity when interacting with search by emoji

* Various fixes

* Theme preview crossfade fixes

* Fix emoji animations positioning on iPad

* Play emoji animations only in personal chats

* Remove test code

* Fixed the flow of messages sent between accounts

+++ Telegram/Watch/WatchCommonWatch/TGBridgeContext.m
@@ -40,7 +40,7 @@ NSString *const TGBridgeContextStartupDataVersion = @"version";
     return dictionary;
 }
 
-- (TGBridgeContext *)updatedWithAuthorized:(bool)authorized peerId:(int32_t)peerId
+- (TGBridgeContext *)updatedWithAuthorized:(bool)authorized peerId:(int64_t)peerId
 {
     TGBridgeContext *context = [[TGBridgeContext alloc] init];
     context->_authorized = authorized;
